Re-enable zluda_inject
