+++ Cargo.toml
@@ -6,7 +6,7 @@ members = [
     "level_zero",
     "spirv_tools-sys",
     "zluda",
-    #"zluda_inject",
+    "zluda_inject",
     "zluda_redirect",
     "ptx",
 ]
