+++ zluda_inject/Cargo.toml
@@ -9,7 +9,6 @@ name = "zluda"
 path = "src/main.rs"
 
 [target.'cfg(windows)'.dependencies]
-zluda_redirect = { path = "../zluda_redirect" }
 winapi = { version = "0.3", features = ["processthreadsapi", "std", "synchapi"] }
 detours-sys = { path = "../detours-sys" }
 clap = "2.33"
